refactor: centralize cross-mixin TYPE_CHECKING stubs in a shared base
Closes #834.

The four Portfolio mixins each carried their own `if TYPE_CHECKING:` block
re-declaring the shared surface they consume (cashposition/prices/aum/
cost_per_unit/cost_bps fields and the cross-mixin returns/profit/
nav_accumulated/turnover/data/assets properties). Those stubs were duplicated
across modules and prone to drift.

Introduce `_portfolio_base._PortfolioMembers`, a type-only base that declares
that surface once under TYPE_CHECKING, and have all four mixins inherit it.
Each mixin keeps only its own mixin-specific stubs (cache slots,
_assert_clean_series, from_cash_position). At runtime the base is an empty
class (the mixins are slot-free already), so instance layout and behaviour are
unchanged.

Net -51/+8 lines across the mixins. Verified: ty + mypy --strict green
(38 files), make test 1210 passed / 5 skipped / 100% coverage, fmt + deptry
clean.
